137 - move otel_to_pv to own module file
