+++ tel2puml/otel_to_pv/sequence_otel.py
@@ -4,11 +4,6 @@ from typing import Any, Generator, Iterable
 from tel2puml.otel_to_pv.otel_to_pv_types import OTelEvent
 from tel2puml.tel2puml_types import PVEvent
 from tel2puml.utils import unix_nano_to_pv_string
-from tel2puml.otel_to_pv.ingest_otel_data import (
-    fetch_data_holder,
-    ingest_data_into_dataholder,
-)
-from tel2puml.otel_to_pv.config import IngestDataConfig, SequenceModelConfig
 
 
 def order_groups_by_start_timestamp(
@@ -308,62 +303,3 @@ def job_ids_to_eventid_to_otelevent_map(
     """
     for job_group in job_id_streams:
         yield {otel_event.event_id: otel_event for otel_event in job_group}
-
-
-def otel_to_pv(
-    config: IngestDataConfig,
-    ingest_data: bool = False,
-    find_unique_graphs: bool = False,
-) -> Generator[
-    tuple[str, Generator[Generator[PVEvent, Any, None], Any, None]], Any, None
-]:
-    """
-    Stream data from data holder and convert to PVEvents. Optional parameters
-    to ingest data as well as find unique graphs within the data set.
-
-    :param config: The configuration for data ingestion and processing.
-    :type config: :class:`IngestDataConfig`
-    :param ingest_data: Flag to indicate whether to load data into the data
-    holder. Defaults to False.
-    :type ingest_data: `bool`, optional
-    :param find_unique_graphs: Flag to indicate whether to find unique graphs
-    within the data holder object. Defaults to False.
-    :type find_unique_graphs: `bool`, optional
-    :return: Generator of tuples of job_name to generator of generators of
-    PVEvents grouped by job_name, then job_id.
-    :rtype: `Generator`[`tuple`[`str`, `Generator`[`Generator`[:class:
-    `PVEvent`, `Any`, `None`], `Any`, `None`]], `Any`, `None`]
-
-    """
-    if ingest_data:
-        data_holder = ingest_data_into_dataholder(config)
-    else:
-        data_holder = fetch_data_holder(config)
-
-    if find_unique_graphs:
-        job_name_to_job_ids_map: dict[str, set[str]] | None = (
-            data_holder.find_unique_graphs()
-        )
-    else:
-        job_name_to_job_ids_map = None
-    job_name_group_streams = data_holder.stream_data(job_name_to_job_ids_map)
-    # get the async event groups from the config
-    sequencer_config = config.get(
-        "sequencer", SequenceModelConfig()
-    )
-    return (
-        (
-            job_name,
-            sequence_otel_job_id_streams(
-                job_id_streams,
-                async_flag=sequencer_config.async_flag,
-                event_to_async_group_map=(
-                    sequencer_config.
-                    async_event_groups.get(
-                        job_name, None
-                    )
-                )
-            ),
-        )
-        for job_name, job_id_streams in job_name_group_streams
-    )
